*: clean useless code && improvement the code struct
