+++ cdc/kv/client_test.go
@@ -17,12 +17,20 @@ import (
 	"context"
 	"fmt"
 	"net"
+	"net/url"
 	"runtime"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"testing"
 	"time"
+
+	"github.com/pingcap/ticdc/pkg/etcd"
+	"go.etcd.io/etcd/clientv3"
+	"go.etcd.io/etcd/embed"
+	"go.etcd.io/etcd/pkg/logutil"
+	"go.uber.org/zap/zapcore"
+	"golang.org/x/sync/errgroup"
 
 	"github.com/golang/protobuf/proto" // nolint:staticcheck
 	"github.com/pingcap/check"
@@ -61,9 +69,43 @@ func Test(t *testing.T) {
 }
 
 type clientSuite struct {
+	e         *embed.Etcd
+	clientURL *url.URL
+	client    etcd.CDCEtcdClient
+	ctx       context.Context
+	cancel    context.CancelFunc
+	errg      *errgroup.Group
 }
 
 var _ = check.Suite(&clientSuite{})
+
+func (s *clientSuite) SetUpTest(c *check.C) {
+	dir := c.MkDir()
+	var err error
+	s.clientURL, s.e, err = etcd.SetupEmbedEtcd(dir)
+	c.Assert(err, check.IsNil)
+	logConfig := logutil.DefaultZapLoggerConfig
+	logConfig.Level = zap.NewAtomicLevelAt(zapcore.ErrorLevel)
+	client, err := clientv3.New(clientv3.Config{
+		Endpoints:   []string{s.clientURL.String()},
+		DialTimeout: 3 * time.Second,
+		LogConfig:   &logConfig,
+	})
+	c.Assert(err, check.IsNil)
+	s.client = etcd.NewCDCEtcdClient(context.TODO(), client)
+	s.ctx, s.cancel = context.WithCancel(context.Background())
+	s.errg = util.HandleErrWithErrGroup(s.ctx, s.e.Err(), func(e error) { c.Log(e) })
+}
+
+func (s *clientSuite) TearDownTest(c *check.C) {
+	s.e.Close()
+	s.cancel()
+	err := s.errg.Wait()
+	if err != nil {
+		c.Errorf("Error group error: %s", err)
+	}
+	s.client.Close() //nolint:errcheck
+}
 
 func (s *clientSuite) TestNewClose(c *check.C) {
 	defer testleak.AfterTest(c)()
@@ -301,8 +343,7 @@ func waitRequestID(c *check.C, allocatedID uint64) {
 	c.Assert(err, check.IsNil)
 }
 
-// Use etcdSuite to workaround the race. See comments of `TestConnArray`.
-func (s *etcdSuite) TestConnectOfflineTiKV(c *check.C) {
+func (s *clientSuite) TestConnectOfflineTiKV(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -401,7 +442,7 @@ func (s *etcdSuite) TestConnectOfflineTiKV(c *check.C) {
 	cancel()
 }
 
-func (s *etcdSuite) TestRecvLargeMessageSize(c *check.C) {
+func (s *clientSuite) TestRecvLargeMessageSize(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -486,7 +527,7 @@ func (s *etcdSuite) TestRecvLargeMessageSize(c *check.C) {
 	cancel()
 }
 
-func (s *etcdSuite) TestHandleError(c *check.C) {
+func (s *clientSuite) TestHandleError(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -658,7 +699,7 @@ consumePreResolvedTs:
 // TestCompatibilityWithSameConn tests kv client returns an error when TiKV returns
 // the Compatibility error. This error only happens when the same connection to
 // TiKV have different versions.
-func (s *etcdSuite) TestCompatibilityWithSameConn(c *check.C) {
+func (s *clientSuite) TestCompatibilityWithSameConn(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -720,7 +761,7 @@ func (s *etcdSuite) TestCompatibilityWithSameConn(c *check.C) {
 	cancel()
 }
 
-func (s *etcdSuite) testHandleFeedEvent(c *check.C) {
+func (s *clientSuite) testHandleFeedEvent(c *check.C) {
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
 	wg := &sync.WaitGroup{}
@@ -1135,12 +1176,12 @@ func (s *etcdSuite) testHandleFeedEvent(c *check.C) {
 	cancel()
 }
 
-func (s *etcdSuite) TestHandleFeedEvent(c *check.C) {
+func (s *clientSuite) TestHandleFeedEvent(c *check.C) {
 	defer testleak.AfterTest(c)()
 	s.testHandleFeedEvent(c)
 }
 
-func (s *etcdSuite) TestHandleFeedEventWithWorkerPool(c *check.C) {
+func (s *clientSuite) TestHandleFeedEventWithWorkerPool(c *check.C) {
 	defer testleak.AfterTest(c)()
 	hwm := regionWorkerHighWatermark
 	lwm := regionWorkerLowWatermark
@@ -1156,7 +1197,7 @@ func (s *etcdSuite) TestHandleFeedEventWithWorkerPool(c *check.C) {
 // TestStreamSendWithError mainly tests the scenario that the `Send` call of a gPRC
 // stream of kv client meets error, and kv client can clean up the broken stream,
 // establish a new one and recover the normal event feed processing.
-func (s *etcdSuite) TestStreamSendWithError(c *check.C) {
+func (s *clientSuite) TestStreamSendWithError(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -1273,7 +1314,7 @@ func (s *etcdSuite) TestStreamSendWithError(c *check.C) {
 	c.Assert(strings.Count(stack, "collectWorkpoolError"), check.Equals, 1)
 }
 
-func (s *etcdSuite) testStreamRecvWithError(c *check.C, failpointStr string) {
+func (s *clientSuite) testStreamRecvWithError(c *check.C, failpointStr string) {
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
 	wg := &sync.WaitGroup{}
@@ -1392,7 +1433,7 @@ eventLoop:
 
 // TestStreamRecvWithErrorAndResolvedGoBack mainly tests the scenario that the `Recv` call of a gPRC
 // stream in kv client meets error, and kv client reconnection with tikv with the current tso
-func (s *etcdSuite) TestStreamRecvWithErrorAndResolvedGoBack(c *check.C) {
+func (s *clientSuite) TestStreamRecvWithErrorAndResolvedGoBack(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	if !util.FailpointBuild {
@@ -1563,7 +1604,7 @@ ReceiveLoop:
 // TestStreamRecvWithErrorNormal mainly tests the scenario that the `Recv` call
 // of a gPRC stream in kv client meets a **logical related** error, and kv client
 // logs the error and re-establish new request.
-func (s *etcdSuite) TestStreamRecvWithErrorNormal(c *check.C) {
+func (s *clientSuite) TestStreamRecvWithErrorNormal(c *check.C) {
 	defer testleak.AfterTest(c)()
 
 	// test client v2
@@ -1581,7 +1622,7 @@ func (s *etcdSuite) TestStreamRecvWithErrorNormal(c *check.C) {
 // TestStreamRecvWithErrorIOEOF mainly tests the scenario that the `Recv` call
 // of a gPRC stream in kv client meets error io.EOF, and kv client logs the error
 // and re-establish new request
-func (s *etcdSuite) TestStreamRecvWithErrorIOEOF(c *check.C) {
+func (s *clientSuite) TestStreamRecvWithErrorIOEOF(c *check.C) {
 	defer testleak.AfterTest(c)()
 
 	// test client v2
@@ -1601,7 +1642,7 @@ func (s *etcdSuite) TestStreamRecvWithErrorIOEOF(c *check.C) {
 // TiCDC will wait 20s and then retry. This is a common scenario when rolling
 // upgrade a cluster and the new version is not compatible with the old version
 // (upgrade TiCDC before TiKV, since upgrade TiKV often takes much longer).
-func (s *etcdSuite) TestIncompatibleTiKV(c *check.C) {
+func (s *clientSuite) TestIncompatibleTiKV(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -1713,7 +1754,7 @@ func (s *etcdSuite) TestIncompatibleTiKV(c *check.C) {
 // TestPendingRegionError tests kv client should return an error when receiving
 // a new subscription (the first event of specific region) but the corresponding
 // region is not found in pending regions.
-func (s *etcdSuite) TestNoPendingRegionError(c *check.C) {
+func (s *clientSuite) TestNoPendingRegionError(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -1803,7 +1844,7 @@ func (s *etcdSuite) TestNoPendingRegionError(c *check.C) {
 }
 
 // TestDropStaleRequest tests kv client should drop an event if its request id is outdated.
-func (s *etcdSuite) TestDropStaleRequest(c *check.C) {
+func (s *clientSuite) TestDropStaleRequest(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -1907,7 +1948,7 @@ func (s *etcdSuite) TestDropStaleRequest(c *check.C) {
 }
 
 // TestResolveLock tests the resolve lock logic in kv client
-func (s *etcdSuite) TestResolveLock(c *check.C) {
+func (s *clientSuite) TestResolveLock(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -2002,7 +2043,7 @@ func (s *etcdSuite) TestResolveLock(c *check.C) {
 	cancel()
 }
 
-func (s *etcdSuite) testEventCommitTsFallback(c *check.C, events []*cdcpb.ChangeDataEvent) {
+func (s *clientSuite) testEventCommitTsFallback(c *check.C, events []*cdcpb.ChangeDataEvent) {
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
 	wg := &sync.WaitGroup{}
@@ -2069,7 +2110,7 @@ func (s *etcdSuite) testEventCommitTsFallback(c *check.C, events []*cdcpb.Change
 }
 
 // TestCommittedFallback tests kv client should panic when receiving a fallback committed event
-func (s *etcdSuite) TestCommittedFallback(c *check.C) {
+func (s *clientSuite) TestCommittedFallback(c *check.C) {
 	defer testleak.AfterTest(c)()
 	events := []*cdcpb.ChangeDataEvent{
 		{Events: []*cdcpb.Event{
@@ -2095,7 +2136,7 @@ func (s *etcdSuite) TestCommittedFallback(c *check.C) {
 }
 
 // TestCommitFallback tests kv client should panic when receiving a fallback commit event
-func (s *etcdSuite) TestCommitFallback(c *check.C) {
+func (s *clientSuite) TestCommitFallback(c *check.C) {
 	defer testleak.AfterTest(c)()
 	events := []*cdcpb.ChangeDataEvent{
 		mockInitializedEvent(3, currentRequestID()),
@@ -2121,7 +2162,7 @@ func (s *etcdSuite) TestCommitFallback(c *check.C) {
 }
 
 // TestDeuplicateRequest tests kv client should panic when meeting a duplicate error
-func (s *etcdSuite) TestDuplicateRequest(c *check.C) {
+func (s *clientSuite) TestDuplicateRequest(c *check.C) {
 	defer testleak.AfterTest(c)()
 	events := []*cdcpb.ChangeDataEvent{
 		{Events: []*cdcpb.Event{
@@ -2142,7 +2183,7 @@ func (s *etcdSuite) TestDuplicateRequest(c *check.C) {
 // testEventAfterFeedStop tests kv client can drop events sent after region feed is stopped
 // TODO: testEventAfterFeedStop is not stable, re-enable it after it is stable
 // nolint:unused
-func (s *etcdSuite) testEventAfterFeedStop(c *check.C) {
+func (s *clientSuite) testEventAfterFeedStop(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -2338,7 +2379,7 @@ func (s *etcdSuite) testEventAfterFeedStop(c *check.C) {
 	cancel()
 }
 
-func (s *etcdSuite) TestOutOfRegionRangeEvent(c *check.C) {
+func (s *clientSuite) TestOutOfRegionRangeEvent(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -2565,7 +2606,7 @@ func (s *clientSuite) TestSingleRegionInfoClone(c *check.C) {
 
 // TestResolveLockNoCandidate tests the resolved ts manager can work normally
 // when no region exceeds reslove lock interval, that is what candidate means.
-func (s *etcdSuite) TestResolveLockNoCandidate(c *check.C) {
+func (s *clientSuite) TestResolveLockNoCandidate(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -2653,7 +2694,7 @@ func (s *etcdSuite) TestResolveLockNoCandidate(c *check.C) {
 // 2. We delay the kv client to re-create a new region request by 500ms via failpoint.
 // 3. Before new region request is fired, simulate kv client `stream.Recv` returns an error, the stream
 //    handler will signal region worker to exit, which will evict all active region states then.
-func (s *etcdSuite) TestFailRegionReentrant(c *check.C) {
+func (s *clientSuite) TestFailRegionReentrant(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -2738,7 +2779,7 @@ func (s *etcdSuite) TestFailRegionReentrant(c *check.C) {
 //    has been deleted in step-3, so it will create new stream but fails because
 //    of unstable TiKV store, at this point, the kv client should handle with the
 //    pending region correctly.
-func (s *etcdSuite) TestClientV1UnlockRangeReentrant(c *check.C) {
+func (s *clientSuite) TestClientV1UnlockRangeReentrant(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 
@@ -2803,7 +2844,7 @@ func (s *etcdSuite) TestClientV1UnlockRangeReentrant(c *check.C) {
 
 // TestClientErrNoPendingRegion has the similar procedure with TestClientV1UnlockRangeReentrant
 // The difference is the delay injected point for region 2
-func (s *etcdSuite) TestClientErrNoPendingRegion(c *check.C) {
+func (s *clientSuite) TestClientErrNoPendingRegion(c *check.C) {
 	defer testleak.AfterTest(c)()
 	clientv2 := enableKVClientV2
 	enableKVClientV2 = false
@@ -2818,7 +2859,7 @@ func (s *etcdSuite) TestClientErrNoPendingRegion(c *check.C) {
 	s.testClientErrNoPendingRegion(c)
 }
 
-func (s *etcdSuite) testClientErrNoPendingRegion(c *check.C) {
+func (s *clientSuite) testClientErrNoPendingRegion(c *check.C) {
 	defer s.TearDownTest(c)
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -2885,7 +2926,7 @@ func (s *etcdSuite) testClientErrNoPendingRegion(c *check.C) {
 }
 
 // TestKVClientForceReconnect force reconnect gRPC stream can work
-func (s *etcdSuite) testKVClientForceReconnect(c *check.C) {
+func (s *clientSuite) testKVClientForceReconnect(c *check.C) {
 	ctx, cancel := context.WithCancel(context.Background())
 	wg := &sync.WaitGroup{}
 
@@ -3023,7 +3064,7 @@ eventLoop:
 	cancel()
 }
 
-func (s *etcdSuite) TestKVClientForceReconnect(c *check.C) {
+func (s *clientSuite) TestKVClientForceReconnect(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 
@@ -3043,7 +3084,7 @@ func (s *etcdSuite) TestKVClientForceReconnect(c *check.C) {
 // TestConcurrentProcessRangeRequest when region range request channel is full,
 // the kv client can process it correctly without deadlock. This is more likely
 // to happen when region split and merge frequently and large stale requests exist.
-func (s *etcdSuite) TestConcurrentProcessRangeRequest(c *check.C) {
+func (s *clientSuite) TestConcurrentProcessRangeRequest(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -3166,7 +3207,7 @@ checkEvent:
 // TestEvTimeUpdate creates a new event feed, send N committed events every 100ms,
 // use failpoint to set reconnect interval to 1s, the last event time of region
 // should be updated correctly and no reconnect is triggered
-func (s *etcdSuite) TestEvTimeUpdate(c *check.C) {
+func (s *clientSuite) TestEvTimeUpdate(c *check.C) {
 	defer testleak.AfterTest(c)()
 
 	defer s.TearDownTest(c)
@@ -3284,7 +3325,7 @@ func (s *etcdSuite) TestEvTimeUpdate(c *check.C) {
 
 // TestRegionWorkerExitWhenIsIdle tests region worker can exit, and cancel gRPC
 // stream automatically when it is idle.
-func (s *etcdSuite) TestRegionWorkerExitWhenIsIdle(c *check.C) {
+func (s *clientSuite) TestRegionWorkerExitWhenIsIdle(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
