+++ cdc/kv/grpc_pool_impl_test.go
@@ -21,11 +21,11 @@ import (
 	"github.com/pingcap/ticdc/pkg/util/testleak"
 )
 
-// Use etcdSuite for some special reasons, the embed etcd uses zap as the only candidate
-// logger and in the logger initializtion it also initializes the grpclog/loggerv2, which
+// Use clientSuite for some special reasons, the embed etcd uses zap as the only candidate
+// logger and in the logger initialization it also initializes the grpclog/loggerv2, which
 // is not a thread-safe operation and it must be called before any gRPC functions
 // ref: https://github.com/grpc/grpc-go/blob/master/grpclog/loggerv2.go#L67-L72
-func (s *etcdSuite) TestConnArray(c *check.C) {
+func (s *clientSuite) TestConnArray(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -59,7 +59,7 @@ func (s *etcdSuite) TestConnArray(c *check.C) {
 	}
 }
 
-func (s *etcdSuite) TestConnArrayRecycle(c *check.C) {
+func (s *clientSuite) TestConnArrayRecycle(c *check.C) {
 	defer testleak.AfterTest(c)()
 	defer s.TearDownTest(c)
 	ctx, cancel := context.WithCancel(context.Background())
