+++ cdc/kv/metrics.go
@@ -87,13 +87,6 @@ var (
 			Help:      "The number of region in one batch resolved ts event",
 			Buckets:   prometheus.ExponentialBuckets(2, 2, 16),
 		}, []string{"capture", "changefeed"})
-	etcdRequestCounter = prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Namespace: "ticdc",
-			Subsystem: "etcd",
-			Name:      "request_count",
-			Help:      "request counter of etcd operation",
-		}, []string{"type", "capture"})
 	grpcPoolStreamGauge = prometheus.NewGaugeVec(
 		prometheus.GaugeOpts{
 			Namespace: "ticdc",
@@ -114,7 +107,6 @@ func InitMetrics(registry *prometheus.Registry) {
 	registry.MustRegister(clientChannelSize)
 	registry.MustRegister(clientRegionTokenSize)
 	registry.MustRegister(batchResolvedEventSize)
-	registry.MustRegister(etcdRequestCounter)
 	registry.MustRegister(grpcPoolStreamGauge)
 
 	// Register client metrics to registry.
