+++ cdc/owner/scheduler.go
@@ -21,6 +21,7 @@ import (
 	"github.com/pingcap/log"
 	"github.com/pingcap/ticdc/cdc/model"
 	cerror "github.com/pingcap/ticdc/pkg/errors"
+	"github.com/pingcap/ticdc/pkg/orchestrator"
 	"go.uber.org/zap"
 )
 
@@ -46,7 +47,7 @@ type moveTableJob struct {
 }
 
 type scheduler struct {
-	state         *model.ChangefeedReactorState
+	state         *orchestrator.ChangefeedReactorState
 	currentTables []model.TableID
 	captures      map[model.CaptureID]*model.CaptureInfo
 
@@ -65,7 +66,7 @@ func newScheduler() *scheduler {
 // Tick is the main function of scheduler. It dispatches tables to captures and handles move-table and rebalance events.
 // Tick returns a bool representing whether the changefeed's state can be updated in this tick.
 // The state can be updated only if all the tables which should be listened to have been dispatched to captures and no operations have been sent to captures in this tick.
-func (s *scheduler) Tick(state *model.ChangefeedReactorState, currentTables []model.TableID, captures map[model.CaptureID]*model.CaptureInfo) (shouldUpdateState bool, err error) {
+func (s *scheduler) Tick(state *orchestrator.ChangefeedReactorState, currentTables []model.TableID, captures map[model.CaptureID]*model.CaptureInfo) (shouldUpdateState bool, err error) {
 	s.state = state
 	s.currentTables = currentTables
 	s.captures = captures
