+++ cdc/processor/manager_test.go
@@ -32,7 +32,7 @@ import (
 
 type managerSuite struct {
 	manager *Manager
-	state   *model.GlobalReactorState
+	state   *orchestrator.GlobalReactorState
 	tester  *orchestrator.ReactorStateTester
 }
 
@@ -48,7 +48,7 @@ func (s *managerSuite) resetSuit(ctx cdcContext.Context, c *check.C) {
 			checkpointTs: replicaInfo.StartTs,
 		}, nil
 	})
-	s.state = model.NewGlobalState().(*model.GlobalReactorState)
+	s.state = orchestrator.NewGlobalState().(*orchestrator.GlobalReactorState)
 	captureInfoBytes, err := ctx.GlobalVars().CaptureInfo.Marshal()
 	c.Assert(err, check.IsNil)
 	s.tester = orchestrator.NewReactorStateTester(c, s.state, map[string]string{
@@ -67,7 +67,7 @@ func (s *managerSuite) TestChangefeed(c *check.C) {
 	c.Assert(err, check.IsNil)
 
 	// an inactive changefeed
-	s.state.Changefeeds["test-changefeed"] = model.NewChangefeedReactorState("test-changefeed")
+	s.state.Changefeeds["test-changefeed"] = orchestrator.NewChangefeedReactorState("test-changefeed")
 	_, err = s.manager.Tick(ctx, s.state)
 	s.tester.MustApplyPatches()
 	c.Assert(err, check.IsNil)
@@ -124,7 +124,7 @@ func (s *managerSuite) TestDebugInfo(c *check.C) {
 	c.Assert(err, check.IsNil)
 
 	// an active changefeed
-	s.state.Changefeeds["test-changefeed"] = model.NewChangefeedReactorState("test-changefeed")
+	s.state.Changefeeds["test-changefeed"] = orchestrator.NewChangefeedReactorState("test-changefeed")
 	s.state.Changefeeds["test-changefeed"].PatchInfo(func(info *model.ChangeFeedInfo) (*model.ChangeFeedInfo, bool, error) {
 		return &model.ChangeFeedInfo{
 			SinkURI:    "blackhole://",
@@ -178,7 +178,7 @@ func (s *managerSuite) TestClose(c *check.C) {
 	c.Assert(err, check.IsNil)
 
 	// an active changefeed
-	s.state.Changefeeds["test-changefeed"] = model.NewChangefeedReactorState("test-changefeed")
+	s.state.Changefeeds["test-changefeed"] = orchestrator.NewChangefeedReactorState("test-changefeed")
 	s.state.Changefeeds["test-changefeed"].PatchInfo(func(info *model.ChangeFeedInfo) (*model.ChangeFeedInfo, bool, error) {
 		return &model.ChangeFeedInfo{
 			SinkURI:    "blackhole://",
