+++ pkg/cmd/cli/cli_changefeed_resume.go
@@ -17,8 +17,9 @@ import (
 	"context"
 	"encoding/json"
 
+	"github.com/pingcap/ticdc/pkg/etcd"
+
 	"github.com/pingcap/ticdc/cdc"
-	"github.com/pingcap/ticdc/cdc/kv"
 	"github.com/pingcap/ticdc/cdc/model"
 	cmdcontext "github.com/pingcap/ticdc/pkg/cmd/context"
 	"github.com/pingcap/ticdc/pkg/cmd/factory"
@@ -29,7 +30,7 @@ import (
 
 // resumeChangefeedOptions defines flags for the `cli changefeed resume` command.
 type resumeChangefeedOptions struct {
-	etcdClient *kv.CDCEtcdClient
+	etcdClient *etcd.CDCEtcdClient
 	pdClient   pd.Client
 
 	credential *security.Credential
