+++ pkg/cmd/cli/cli_processor_list.go
@@ -14,16 +14,16 @@
 package cli
 
 import (
-	"github.com/pingcap/ticdc/cdc/kv"
 	"github.com/pingcap/ticdc/pkg/cmd/context"
 	"github.com/pingcap/ticdc/pkg/cmd/factory"
 	"github.com/pingcap/ticdc/pkg/cmd/util"
+	"github.com/pingcap/ticdc/pkg/etcd"
 	"github.com/spf13/cobra"
 )
 
 // listProcessorOptions defines flags for the `cli processor list` command.
 type listProcessorOptions struct {
-	etcdClient *kv.CDCEtcdClient
+	etcdClient *etcd.CDCEtcdClient
 }
 
 // newListProcessorOptions creates new listProcessorOptions for the `cli processor list` command.
