+++ pkg/cmd/util/helper.go
@@ -22,10 +22,11 @@ import (
 	"strings"
 	"syscall"
 
+	"github.com/pingcap/ticdc/pkg/etcd"
+
 	"github.com/BurntSushi/toml"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/log"
-	"github.com/pingcap/ticdc/cdc/kv"
 	cmdconetxt "github.com/pingcap/ticdc/pkg/cmd/context"
 	"github.com/pingcap/ticdc/pkg/logutil"
 	"github.com/pingcap/ticdc/pkg/version"
@@ -151,7 +152,7 @@ func JSONPrint(cmd *cobra.Command, v interface{}) error {
 // VerifyAndGetTiCDCClusterVersion verifies and gets the version of ticdc.
 // If it is an incompatible version, an error is returned.
 func VerifyAndGetTiCDCClusterVersion(
-	ctx context.Context, cdcEtcdCli *kv.CDCEtcdClient,
+	ctx context.Context, cdcEtcdCli *etcd.CDCEtcdClient,
 ) (version.TiCDCClusterVersion, error) {
 	_, captureInfos, err := cdcEtcdCli.GetCaptures(ctx)
 	if err != nil {
