+++ tests/move_table/main.go
@@ -19,6 +19,7 @@ import (
 	"context"
 	"flag"
 	"fmt"
+	"github.com/pingcap/ticdc/pkg/etcd"
 	"io/ioutil"
 	"net/http"
 	"strings"
@@ -26,7 +27,6 @@ import (
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/log"
-	"github.com/pingcap/ticdc/cdc/kv"
 	cerrors "github.com/pingcap/ticdc/pkg/errors"
 	"github.com/pingcap/ticdc/pkg/retry"
 	"go.etcd.io/etcd/clientv3"
@@ -143,7 +143,7 @@ type tableInfo struct {
 type cluster struct {
 	ownerAddr  string
 	captures   map[string][]*tableInfo
-	cdcEtcdCli kv.CDCEtcdClient
+	cdcEtcdCli etcd.CDCEtcdClient
 }
 
 func newCluster(ctx context.Context, pd string) (*cluster, error) {
@@ -177,7 +177,7 @@ func newCluster(ctx context.Context, pd string) (*cluster, error) {
 	ret := &cluster{
 		ownerAddr:  "",
 		captures:   nil,
-		cdcEtcdCli: kv.NewCDCEtcdClient(ctx, etcdCli),
+		cdcEtcdCli: etcd.NewCDCEtcdClient(ctx, etcdCli),
 	}
 
 	log.Info("new cluster initialized")
@@ -186,7 +186,7 @@ func newCluster(ctx context.Context, pd string) (*cluster, error) {
 }
 
 func (c *cluster) refreshInfo(ctx context.Context) error {
-	ownerID, err := c.cdcEtcdCli.GetOwnerID(ctx, kv.CaptureOwnerKey)
+	ownerID, err := c.cdcEtcdCli.GetOwnerID(ctx, etcd.CaptureOwnerKey)
 	if err != nil {
 		return errors.Trace(err)
 	}
